requirements-dev.txt
unittest
moto
requests-mock
boto3
pyspark
